// packages/shared/src/media-library.generated.ts
export const MEDIA_LIBRARY_GENERATED: GeneratedMediaLibraryItem[] = [
  {
    "id": "food-beverage-activity-square",
    "categoryGroup": "food-beverage",
    "label": "Activity Square",
    "path": "/library/food-beverage/activity-square.svg"
  },
// ...
];